Progress: Mostly done with adding the Address model.
I'd like to use enums that map actual strings to countries instead of the hash keys being numbers; no one knows what those are when they read a DB. It's not so straightforward to use the collection_select helper when you aren't pulling from a model. The model has 'methods' which are named the same as its keys for properties (i.e., City.id is a method that gets you the 'id') and maybe my Ruby is rusty, but they keys of a Hash aren't methods. I'll try to come up with something else.

I fixed issues with Minitest (DB priveleges) and my fixtures are configured so that entity relationships are respected, a test is initiated with a Person that has Address, Phone, Email, (has_many) and it shows up in the forms

CRUD is working fine for all of these as well.